refactor: page route
