+++ src/modules/contents/attachments/composables/use-attachment.ts
@@ -69,6 +69,8 @@ export function useAttachmentControl(filterOptions?: {
           group?.value?.metadata.name,
           user?.value?.metadata.name,
           attachments.value.size,
+          // TODO
+          // @ts-ignore
           attachments.value.page,
           [],
           []
