+++ src/modules/contents/pages/SinglePageEditor.vue
@@ -1,0 +1,5 @@
+<script lang="ts" setup></script>
+
+<template>
+  <div>Page Editor</div>
+</template>
